+++ src/QRMsep.h
@@ -19,7 +19,7 @@
 
 /* Contact: Alexander J. McNeil:  mcneil@math.ethz.ch */
 
-void SEprocExciteFunc(long *n, double *times, long *nmarks, double *marktimes, double *marks, double *beta, long *model, double *result);
+void SEprocExciteFunc(int *n, double *times, int *nmarks, double *marktimes, double *marks, double *beta, int *model, double *result);
 
 double contribH(double s, double y, double gammaval, double deltaval);
